Moved creating LevelCompleteDialog to LevelScreen
Refactoring GameEngine and subclass CatchDaStars to only contain game logic and
let LevelScreen handle UI management like creating Dialogs.
Added new class Score to contain score items that must be displayed in
LevelCompleteDialog. This was needed to allow the game to pass the score items
to LevelScreen in order to setup the LevelCompleteDialog.